usbwishbonebridge: work with additional endpoints
Due to the way USB interleves packets, we were hitting conditions where
the host would send a SETUP0, followed by DATA, followed by IN1.  We
would respond to this IN1 as if it were an IN0, causing data to become
very corrupted.

Only respond if the endpoint number is 0.  Send NAK otherwise, as long
as we're in a debug condition.

Signed-off-by: Sean Cross <sean@xobs.io>