sm: transfer: remove gpio for pullup
The pullup is part of a GPIO that is used to enable/disable the USB
pullup.

However, a GPIO already exists in the epfifo block, and we don't want a
GPIO when we use the dummy block.  Therefore this GPIO is not only
redundant, but is actively harmful.

Remove this, and rely on the GPIO present in epfifo instead.

Signed-off-by: Sean Cross <sean@xobs.io>