eptri: refactor based on tinyusb work
After working with tinyusb, several design defects were discovered in
the overall design of eptri.  The biggest of these was how we would
specially handle SETUP transactions, which meant that we couldn't be
interrupted mid-transaction.

This reworks many of the registers.

Barring any major discoveries, this is the register set we will ship
with.

Signed-off-by: Sean Cross <sean@xobs.io>