eptri: inhandler: default to DATA0 on non-EP0 endpoints
Previously, we were defaulting to DATA1 on all endpoints.  Due to how
the USB spec works, only EP0 starts on DATA1 -- all other endpoints
start on DATA0.

(Actually, all endpoints start at DATA0, it's just that EP0 is the only
endpoint where OUT and IN packets are correlated, so if the host sends
SETUP, it is always followed by DATA0, we must respond with DATA1.)

Fix the defaults so all other endpoints default to DATA0.

Signed-off-by: Sean Cross <sean@xobs.io>