eptri: use EventSourceProcess to signal usb reset
USB resets may be prolonged -- we detect them after 2uS, but they can be
as long as 2mS.  Previously, we were using an EventSourcePulse for this,
which meant we were generating an endless stream of interrupts during a
reset condition.

Use an EventSourceProcess with an inverted signal to turn the
level-triggered USB reset event into an edge-triggered one.

Signed-off-by: Sean Cross <sean@xobs.io>